pyproject.toml
[project]
name = "doc-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.0",
    "filelock>=3.18.0",
    "openai>=1.95.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "uvicorn>=0.35.0",
]
